@@ -16,14 +16,15 @@
     "dc": "^4.2.7",
     "firebase": "^10.9.0",
     "leaflet": "^1.9.4",
-    "next": "14.1.3",
+    "next": "^14.2.3",
     "nodemailer": "^6.9.13",
-    "react": "^18",
+    "react": "^18.3.1",
     "react-datepicker": "^6.6.0",
-    "react-dom": "^18",
+    "react-dom": "^18.3.1",
     "react-icons": "^5.0.1",
     "react-leaflet": "^4.2.1",
     "react-slick": "^0.30.2",
+    "resend": "^3.2.0",
     "slick-carousel": "^1.8.1",
     "swiper": "^11.1.0",
     "swiper-react": "^1.0.0"
