@@ -1,65 +1,64 @@
-import { NextResponse } from "next/server";
+import { NextResponse } from "next/server";  
 import { db, collection, addDoc, getDocs } from "../../../../../firebase";
-const nodemailer = require("nodemailer");
+//import nodemailer from "nodemailer";
+import { Resend } from 'resend';
+
+const resend = new Resend(process.env.NEXT_PUBLIC_RESEND_API_KEY);
+
+
 async function folioTicket(errorSeleccionado, rutaError) {
+ 
   const refTickets = collection(db, "tickets");
-  const ticketsSnapShot = await getDocs(refTickets);
-  ticketsSnapShot.forEach((doc) => {
-    const tickets = doc.data();
-    console.log(tickets);
+  const querySnapshot = await getDocs(refTickets);
+  let maxNumeroFolio = 0;
+
+  // Recorrer los documentos para encontrar el máximo número de folio del tipo de error seleccionado
+  querySnapshot.forEach((doc) => {
+    const data = doc.data();
+    const folio = data.folio;
+
+    // Verificar si el folio coincide con el tipo de error seleccionado y extraer el número de folio
+    if (folio.startsWith(errorSeleccionado + "-")) {
+      const numeroFolio = parseInt(folio.split("-")[1]);
+      if (numeroFolio > maxNumeroFolio) {
+        maxNumeroFolio = numeroFolio;
+      }
+    }
   });
-  console.log(errorSeleccionado, " ", rutaError);
-  const fol = `${errorSeleccionado}-${rutaError}`;
-  console.log(fol);
+
+  // Incrementar el número de folio en 1
+  const nuevoNumeroFolio = maxNumeroFolio + 1;
+
+  // Construir el nuevo folio
+  const nuevoFolio = `${errorSeleccionado}-${nuevoNumeroFolio}`;
+
+  return nuevoFolio;
 }
 
 function prioridad(errorSeleccionado) {
   let priori = "";
   switch (errorSeleccionado) {
-    case "S001" || "S002" || "0000":
+    case "S001":
+    case "S002":
+    case "0000":
       priori = "PRIORIDAD ALTA";
       break;
-    case "D001" ||
-      "D002" ||
-      "M001" ||
-      "M002" ||
-      "R001" ||
-      "R002" ||
-      "R003" ||
-      "P001" ||
-      "P002" ||
-      "T001":
+    case "D001":
+    case "D002":
+    case "M001":
+    case "M002":
+    case "R001":
+    case "R002":
+    case "R003":
+    case "P001":
+    case "P002":
+    case "T001":
       priori = "PRIORIDAD MEDIA";
       break;
+    default:
+      priori = "PRIORIDAD DESCONOCIDA";
   }
-}
-
-const transporter = nodemailer.createTransport({
-  host: 'smtp.gmail.com',
-  port: 465,
-  secure: true,
-
-  auth: {
-    user: 'somos.gemma.01@gmail.com',
-    pass: 'rje3vw$x4hGVhcw$h8eeukGSh5$C6@o6W5NH'
-  }
-});
-async function enviarCorreo(destinatario, folio) {
-  try {
-    // Configura el contenido del correo electrónico
-    const mailOptions = {
-      from: 'somos.gemma.01@gmail.com',
-      to: destinatario,
-      subject: 'Confirmación de recepción de ticket',
-      text: `Se ha recibido su ticket con el folio: 1.`
-    };
-
-    // Envía el correo electrónico
-    const info = await transporter.sendMail(mailOptions);
-    console.log("Correo enviado:", info.messageId);
-  } catch (error) {
-    console.error("Error al enviar el correo electrónico:", error);
-  }
+  return priori; // Devuelve el string con la prioridad
 }
 
 export async function POST(req, { params }) {
@@ -74,10 +73,16 @@ export async function POST(req, { params }) {
       navegador,
       rutaError,
       descripcionProblema,
+      correoA,
+      nombre,
       url,
+      area
+
     ] = params.Soporte;
     const rutitaD = decodeURIComponent(rutaError);
     const urlsitaD = decodeURIComponent(url);
+    const folio = await folioTicket(errorSeleccionado, rutaError);
+    const priori = prioridad(errorSeleccionado);
     // Validar los datos si es necesario
     console.log(
       errorSeleccionado,
@@ -90,12 +95,17 @@ export async function POST(req, { params }) {
       " ",
       descripcionProblema,
       " ",
-      urlsitaD
-    );
-    // Guardar los datos en la base de datos
+      urlsitaD, " ", priori, " ", folio
+    ); 
+    console.log(resend)
+    resend.emails.send({
+      from: 'onboarding@resend.dev',
+      to: correoA,
+      subject: 'Confirmación de recepción de ticket',
+      html: `Se ha recibido su ticket con el folio: ${folio}.`
+    });
+    
 
-    const folio = folioTicket(errorSeleccionado, rutaError);
-    const priori = prioridad(errorSeleccionado);
 
     const docRef = await addDoc(collection(db, 'tickets'), {
         folio,
@@ -105,12 +115,16 @@ export async function POST(req, { params }) {
         navegador,
         rutitaD,
         descripcionProblema,
-        timestamp: new Date(),
+        fechaDeEnvio: new Date(),
         urlsitaD,
-      }); 
-      await enviarCorreo('melyssagabrielag@gmail.com', folio);
+        nombre,
+        correoA,
+        area
+      });
+
+  console.log(docRef)
     // Enviar una respuesta de éxito
-    return NextResponse.json(/* docRef */ "s");
+    return NextResponse.json(docRef );
   } catch (error) {
     console.error("Error al obtener reportes:", error);
     return NextResponse.error("Error al obtener reportes", { status: 500 });
