@@ -1,10 +1,34 @@
 "use client";
-import React, { useState, useRef } from 'react';
+import React, { useState, useRef, useEffect, useContext } from "react";
+import { getStorage, ref, uploadBytes, getDownloadURL } from "firebase/storage";
+import { useAuthUser } from "../../../../../hooks/UseAuthUser";
+import { app, auth, db } from "../../../../../firebase";
+import { useRouter } from "next/navigation";
+import AuthContext from "../../../../../context/AuthContext";
 import RutaProtegida from "@/components/RutaProtegida";
+import "./Reportes.css";
 import "./Soporte.css";
 
 
 function Soporte() {
+  useAuthUser();
+  const [fecha, setFecha] = useState('');
+  const [rutaError, setRutaError] = useState('');
+  const textareaRef = useRef(null);
+  const router = useRouter();
+  const { isLogged } = useContext(AuthContext);
+  const [userData, setUserData] = useState({});
+  const [errorSeleccionado, setErrorSeleccionado] = useState("S001");
+  const [sistemaOperativo, setSistemaOperativo] = useState(
+    "No se ha seleccionado un sistema operativo"
+  );
+  const [asignarTarea, setAsignarTarea] = useState("-")
+  const [navegador, setNavegador] = useState(
+    "No se ha seleccionado un navegador"
+  );
+  const [selectedRutaError, setSelectedRutaError] = useState("/NoEspecificado");
+  const [foto, setFoto] = useState("");
+  const [descripcionProblema, setDescripcionProblema] = useState("Sin descripcion");
 
   const [mostrarDetalle1, setMostrarDetalle1] = useState(false);
   const [mostrarDetalle2, setMostrarDetalle2] = useState(false);
@@ -44,6 +68,46 @@ function Soporte() {
     return mostrarDetalle4 ? 'https://i.postimg.cc/wB2S0JqV/triangulo-1.png' : 'https://i.postimg.cc/Z5XxTn6Y/triangulo.png';
   };
 
+  useEffect(() => {
+    if (typeof window !== "undefined") {
+      const unsubscribe = auth.onAuthStateChanged((user) => {
+        if (user) {
+          const uid = user.uid;
+          fetchData(uid);
+        } else {
+          router.push("/reportes");
+        }
+      });
+      return () => unsubscribe();
+    }
+
+    async function fetchData(uid) {
+      try {
+        const userResponse = await fetch(`/api/Usuario/${uid}`);
+        if (!userResponse.ok) {
+          throw new Error("Failed to fetch user data");
+        }
+        const userDatas = await userResponse.json();
+
+        setUserData(userDatas);
+      } catch (error) {
+        console.error("Error fetching data:", error);
+      }
+    }
+  }, []);
+  const catalogoRutaErrores = [
+    { ruta: "/Cuenta/Administrador", modulo: "Inicio de Sesión" },
+    { ruta: "/Administrador/Dashboard", modulo: "Dashboard" },
+    { ruta: "/Administrador/Mapa", modulo: "Mapa" },
+    { ruta: "/Administrador/NuevoAdmin", modulo: "Nuevo Administrador" },
+    { ruta: "/Administrador/Reportes", modulo: "Reportes" },
+    { ruta: "/Administrador/Papelera", modulo: "Reportes" },
+    { ruta: "Otros", modulo: "Otra opción" },
+  ];
+  /*
+  Inicio de sesion , Nuevo Administrador, Otros -> ALTA
+  DASH, MAPA, REPORTES, PAPELERA -> Media
+  */
 
   // Catálogo de errores
   const catalogoErrores = [
@@ -58,7 +122,7 @@ function Soporte() {
     { clave: "R003", nombre: "Error al Mover reportes a la papelera" },
     { clave: "P001", nombre: "Error al Visualizar reportes en la papelera" },
     { clave: "P002", nombre: "Error al Eliminar reportes de la papelera" },
-    { clave: "S001", nombre: "Error al Enviar Ticket" },
+    { clave: "T001", nombre: "Error al Enviar Ticket" },
     { clave: "0000", nombre: "Otro: (Especificar en Descripcion)" },
   ];
 
@@ -86,19 +150,10 @@ function Soporte() {
     "Opera",
     "Internet Explorer",
     "Samsung Internet",
-    "Otro"
+    "Otro",
   ];
 
 
-
-  const [fecha, setFecha] = useState('');
-  const [errorSeleccionado, setErrorSeleccionado] = useState('No se ha seleccionado un error');
-  const [sistemaOperativo, setSistemaOperativo] = useState('No se ha seleccionado un sistema operativo');
-  const [navegador, setNavegador] = useState('No se ha seleccionado un navegador');
-  const [rutaError, setRutaError] = useState('');
-  const [foto, setFoto] = useState('');
-  const [descripcionProblema, setDescripcionProblema] = useState('');
-  const textareaRef = useRef(null);
 
   // Obtener fecha actual al cargar el componente
   /*
@@ -118,6 +173,7 @@ function Soporte() {
 
   // Funciones para manejar los cambios en el select de errores y sistemas operativos
 
+
   const handleError = (e) => {
     const selectedErr = e.target.value;
     setErrorSeleccionado(selectedErr);
@@ -128,14 +184,15 @@ function Soporte() {
     const selectedSO = e.target.value;
     setSistemaOperativo(selectedSO);
     console.log(selectedSO);
-    setSistemaOperativo(selectedSO);
-    console.log(selectedSO);
   };
 
+  const handleAsignarTarea = (e) => {
+    const asignar = e.target.value
+    setAsignarTarea(asignar)
+    console.log(asignar)
+  }
   const handleNavegador = (e) => {
     const selectedNavegador = e.target.value;
-    setNavegador(selectedNavegador);
-    console.log(selectedNavegador);
     setNavegador(selectedNavegador);
     console.log(selectedNavegador);
   };
@@ -183,38 +240,52 @@ function Soporte() {
   const handleSubmit = async (e) => {
     e.preventDefault();
 
-    try {
-      const response = await fetch('/api/Soporte', {
-        method: 'POST',
-        headers: {
-          'Content-Type': 'application/json'
-        },
-        body: JSON.stringify({
-          errorSeleccionado,
-          sistemaOperativo,
-          navegador,
-          rutaError,
-          descripcionProblema
-        })
-      });
+    const correoA = userData.correo;
+    const nombre = userData.nombre;
+    const area = asignarTarea
+    const url = await handleFileUpload();
+    const parametros = {
+      errorSeleccionado: errorSeleccionado,
+      sistemaOperativo: sistemaOperativo,
+      navegador: navegador,
+      selectedRutaError: encodeURIComponent(selectedRutaError),
+      descripcionProblema: descripcionProblema,
+      correoA: correoA,
+      nombre: nombre,
+      url: encodeURIComponent(url),
+      area: area
+    };
+    let res = prompt("A la hora de levantar el ticket, vamos a recuperar su información para darle seguimiento a su ticket, desea aceptar?")
+    if (res === "SI") {
+      try {
+        const response = await fetch(
+          `http://localhost:3001/api/Ticket/${errorSeleccionado}/${sistemaOperativo}/${navegador}/${encodeURIComponent(
+            selectedRutaError
+          )}/${descripcionProblema}/${correoA}/${nombre}/${encodeURIComponent(url)}/${area}`,
+          {
+            method: "POST",
+            headers: {
+              "Content-Type": "application/json",
+            },
+            body: JSON.stringify(parametros),
+          }
+        );
 
-      if (response.ok) {
-        console.log("Formulario enviado con éxito");
-      } else {
-        console.error("Error al enviar el formulario:", response.status);
+        if (response.ok) {
+          console.log("Formulario enviado con éxito");
+        } else {
+          console.error("Error al enviar el formulario:", response.status);
+        }
+      } catch (error) {
+        console.error("Error al enviar el formulario:", error);
       }
-    } catch (error) {
-      console.error("Error al enviar el formulario:", error);
+    } else {
+      alert("NO SE LEVANTARA SU TICKET")
     }
+
   };
 
-  // Catálogo de errores, SO, versión de navegador, descripción del problema, ruta donde se encontró el error, evidencia fotográfica del problema, 
-  // Internamente se obtiene fecha, prioridad con base al error seleccionado, número del ticket, su correo
-
-  //<input type="text" value={fecha} readOnly />
-
   return (
-
     <RutaProtegida>
       <div className="bodySoporte">
         <div className="containerSoporte">
@@ -309,8 +380,87 @@ function Soporte() {
               <h2 id='titulo_sp'>Formulario de Soporte Técnico 👷</h2>
 
               <form onSubmit={handleSubmit}>
+                <label>Seleccione el error:</label>
+                <select value={errorSeleccionado} onChange={handleError}>
+                  <option>Tipo de Error</option>
+                  {catalogoErrores.map((errorSeleccionado, index) => (
+                    <option key={index} value={errorSeleccionado.clave}>
+                      {`${errorSeleccionado.nombre}`}
+                    </option>
+                  ))}
+                </select>
+                <br />
+                <br />
+                <br />
+                <label>Módulo donde se encontró el error: </label>
+                <select value={selectedRutaError} onChange={handleRutaError}>
+                  <option>Módulo del Error</option>
+                  {catalogoRutaErrores.map((errorOption, index) => (
+                    <option key={index} value={errorOption.ruta}>
+                      {`${errorOption.modulo}`}
+                    </option>
+                  ))}
+                </select>
+                <br />
+                <br />
+                <br />
+                <label>Carácter de error</label>
+                <select value={asignarTarea} onChange={handleAsignarTarea}>
+                  <option >Escoger carácter de error</option>
+                  <option value="backend">Funcionalidad</option>
+                  <option value="frontend">Diseño</option>
+                </select>
+                <label>Seleccione su sistema operativo: </label>
+                <select value={sistemaOperativo} onChange={handleSO}>
+                  <option value="">Seleccionar</option>
+                  {catalogoSistemaOperativo.map((sistema, index) => (
+                    <option key={index} value={sistema}>
+                      {`${sistema}`}
+                    </option>
+                  ))}
+                </select>
+                <br />
+                <br />
+                <br />
 
-                {/* 
+                <label>Seleccione su navegador: </label>
+                <select value={navegador} onChange={handleNavegador}>
+                  <option value="">Seleccionar</option>
+                  {catalogoNavegadores.map((navegador, index) => (
+                    <option key={index} value={navegador}>
+                      {`${navegador}`}
+                    </option>
+                  ))}
+                </select>
+                <br />
+                <br />
+                <br />
+
+                <label>Adjuntar fotografía del problema: </label>
+                <input type="file" accept="image/*" onChange={handleFoto} />
+                <br />
+                <br />
+                <br />
+
+                <label>Descripción del problema: </label>
+                <textarea
+                  value={descripcionProblema}
+                  onChange={handleDescripcionProblema}
+                  rows="4"
+                  cols="50"
+                />
+                <br />
+                <br />
+                <br />
+
+                <button type="submit" id="submit">
+                  Enviar
+                </button>
+              </form>
+              {/*
+              <form onSubmit={handleSubmit}>
+
+                
                 <label>Fecha:</label>
                 <input type="text" defaultValue={fecha} readOnly/>
                 <br /><br />
@@ -320,7 +470,7 @@ function Soporte() {
 
                 <h5>Obtener correo del usuario acá</h5>
                 <br />
-              */}
+              
 
                 <label>Seleccione el error ❌:</label>
                 <br />
@@ -408,15 +558,16 @@ function Soporte() {
                 <button type="submit" id="submit">Enviar</button>
 
               </form>
+              */}
+
               <br /><br />
             </div>
           </div>
+
         </div>
       </div>
     </RutaProtegida>
-
   );
 }
 
 export default Soporte;
-
