@@ -1,30 +1,30 @@
-"use client"
-import React, { useState, useEffect, useContext } from 'react';
-import Link from 'next/link';
-import './Navbar.css';
+"use client";
+import React, { useState, useEffect, useContext } from "react";
+import Link from "next/link";
+import "./Navbar.css";
 import AuthContext from "../../context/AuthContext";
 import { useAuthUser } from "../../hooks/UseAuthUser";
-import { auth } from '../../firebase';
-import { signOut } from 'firebase/auth';
-import { useRouter } from 'next/navigation';
-import ConfirmationAlert from '@/components/ConfirmationAlert'; // Importa el componente de alerta de confirmación
+import { auth } from "../../firebase";
+import { signOut } from "firebase/auth";
+import { useRouter } from "next/navigation";
+import ConfirmationAlert from "@/components/ConfirmationAlert"; // Importa el componente de alerta de confirmación
 
 function Navbar() {
   useAuthUser();
-  const { isLogged, isAdmin } = useContext(AuthContext);
+  const { isLogged, isAdmin, isDev } = useContext(AuthContext);
   const [menuActive, setMenuActive] = useState(false);
   const [showMenuIcon, setShowMenuIcon] = useState(false);
   const [showConfirmation, setShowConfirmation] = useState(false); // Estado para mostrar la alerta de confirmación
-  const router = useRouter()
+  const router = useRouter();
   useEffect(() => {
     const handleResize = () => {
       setShowMenuIcon(window.innerWidth <= 800);
     };
 
     handleResize(); // Call on initial render
-    window.addEventListener('resize', handleResize);
+    window.addEventListener("resize", handleResize);
 
-    return () => window.removeEventListener('resize', handleResize);
+    return () => window.removeEventListener("resize", handleResize);
   }, []);
 
   const toggleMenu = () => {
@@ -35,11 +35,11 @@ function Navbar() {
     // Función para manejar el cierre de sesión después de la confirmación
     signOut(auth)
       .then(() => {
-        console.log('Cierre de sesión exitoso');
-        router.push("/Cuenta"); 
+        console.log("Cierre de sesión exitoso");
+        router.push("/Cuenta");
       })
       .catch((error) => {
-        console.error('Error al cerrar sesión:', error);
+        console.error("Error al cerrar sesión:", error);
       });
   };
 
@@ -55,7 +55,11 @@ function Navbar() {
   };
 
   return (
-    <div className={`navBar ${isAdmin ? 'admin' : ''} ${menuActive ? 'showMenu' : ''}`}>
+    <div
+      className={`navBar ${isAdmin ? "admin" : ""} ${
+        menuActive ? "showMenu" : ""
+      }`}
+    >
       <Link href="/" className="bachecito26">
         <img
           src="https://i.postimg.cc/T3NQg97V/Logo.png"
@@ -65,8 +69,11 @@ function Navbar() {
         BACHECITO 26
       </Link>
 
-      <div className={`menuIcon ${showMenuIcon ? 'span-anime' : 'hidden'}`} onClick={toggleMenu}>
-        <div className={`menu-icon ${menuActive ? '' : ''}`}>
+      <div
+        className={`menuIcon ${showMenuIcon ? "span-anime" : "hidden"}`}
+        onClick={toggleMenu}
+      >
+        <div className={`menu-icon ${menuActive ? "" : ""}`}>
           <span></span>
           <span></span>
           <span></span>
@@ -77,55 +84,141 @@ function Navbar() {
           <>
             {isAdmin ? (
               <>
-                <Link href="/Cuenta/Administrador/Dashboard" className="opc-admin"><img src="https://i.postimg.cc/3JkMwkG1/estadisticas-1.png" alt="estadisticas" /><span className='hover-text'>Dashboard</span></Link>
-                <Link href="/Cuenta/Administrador/Reportes" className="opc-admin"><img src="https://i.postimg.cc/6QLgPnsW/encuesta-h-1.png" alt="reportes" /><span className='hover-text'>Reportes</span></Link>
-                <Link href="/Cuenta/Administrador/Mapa" className="opc-admin"><img src="https://i.postimg.cc/QMrvSSyY/marcador-de-mapa-1.png" alt="mapa" /><span className='hover-text'>Mapa</span></Link>
-                <Link href="/Cuenta/Administrador/NuevoAdmin" className="opc-admin"><img src="https://i.postimg.cc/02FTK9ds/agregar-usuario-1.png" alt="administrador nuevo" /><span className='hover-text'>Administradores</span></Link>
-                <Link href="/Cuenta/Administrador/Soporte" className="opc-admin"><img src="https://i.postimg.cc/HkMfQFB7/constructor-1.png" alt="soporte" /><span className='hover-text'>Soporte</span></Link>
-                <Link href="#" className="opc-admin"><img src="https://i.postimg.cc/qRJSHq08/salida-2.png" alt="salir" onClick={CerrarSesion} /><span className='hover-text'>Salir</span></Link>
+                <Link
+                  href="/Cuenta/Administrador/Dashboard"
+                  className="opc-admin"
+                >
+                  <img
+                    src="https://i.postimg.cc/3JkMwkG1/estadisticas-1.png"
+                    alt="estadisticas"
+                  />
+                  <span className="hover-text">Dashboard</span>
+                </Link>
+                <Link
+                  href="/Cuenta/Administrador/Reportes"
+                  className="opc-admin"
+                >
+                  <img
+                    src="https://i.postimg.cc/6QLgPnsW/encuesta-h-1.png"
+                    alt="reportes"
+                  />
+                  <span className="hover-text">Reportes</span>
+                </Link>
+                <Link href="/Cuenta/Administrador/Mapa" className="opc-admin">
+                  <img
+                    src="https://i.postimg.cc/QMrvSSyY/marcador-de-mapa-1.png"
+                    alt="mapa"
+                  />
+                  <span className="hover-text">Mapa</span>
+                </Link>
+                <Link
+                  href="/Cuenta/Administrador/NuevoAdmin"
+                  className="opc-admin"
+                >
+                  <img
+                    src="https://i.postimg.cc/02FTK9ds/agregar-usuario-1.png"
+                    alt="administrador nuevo"
+                  />
+                  <span className="hover-text">Administradores</span>
+                </Link>
+                <Link
+                  href="/Cuenta/Administrador/Soporte"
+                  className="opc-admin"
+                >
+                  <img
+                    src="https://i.postimg.cc/HkMfQFB7/constructor-1.png"
+                    alt="soporte"
+                  />
+                  <span className="hover-text">Soporte</span>
+                </Link>
+                <Link href="#" className="opc-admin">
+                  <img
+                    src="https://i.postimg.cc/qRJSHq08/salida-2.png"
+                    alt="salir"
+                    onClick={CerrarSesion}
+                  />
+                  <span className="hover-text">Salir</span>
+                </Link>
               </>
             ) : (
               <>
-                <Link href="/Cuenta/Usuario/Perfil" className="opc btn--white prueba">
-                  <span>Perfil →</span>
-                </Link>
-                <Link href="/Cuenta/Usuario/Favoritos" className="opc">
-                  Baches Guardados
-                </Link>
-                <Link href="/Cuenta/Usuario/Estadisticas" className="opc">
-                  Estadísticas
-                </Link>
-                <Link href="/Cuenta/Usuario/Reportes" className="opc">
-                  Reportes
-                </Link>
+                {isDev ? (
+                  <>
+                  <Link
+                      href="/Cuenta/Desarrolladores"
+                      className="opc"
+                    >
+                      Desarrolladores
+                    </Link>
+                    <Link
+                      href="/Cuenta/Desarrolladores/NuevoDev"
+                      className="opc"
+                    >
+                      Nuevo Desarrollador
+                    </Link>
+                    <Link
+                      href="/Cuenta/Desarrolladores/Tickets"
+                      className="opc"
+                    >
+                      Tickets
+                    </Link>
+                    <Link href="#" className="opc-admin"> 
+                      <img
+                        src="https://i.postimg.cc/qRJSHq08/salida-2.png"
+                        alt="salir"
+                        onClick={CerrarSesion}
+                      />
+                      <span className="hover-text">Salir</span>
+                    </Link>
+                  </>
+                ) : (
+                  <>
+                    <Link
+                      href="/Cuenta/Usuario/Perfil"
+                      className="opc btn--white prueba"
+                    >
+                      <span>Perfil →</span>
+                    </Link>
+                    <Link href="/Cuenta/Usuario/Favoritos" className="opc">
+                      Baches Guardados
+                    </Link>
+                    <Link href="/Cuenta/Usuario/Estadisticas" className="opc">
+                      Estadísticas
+                    </Link>
+                    <Link href="/Cuenta/Usuario/Reportes" className="opc">
+                      Reportes
+                    </Link>
+                  </>
+                )}
               </>
             )}
           </>
         )}
         {!isLogged && (
           <>
-           <Link href="/Sobre_Nosotros" className="opc">
-             Sobre Nosotros
-           </Link>
-           <Link href="/Reportes" className="opc">
-             Reportes
-           </Link>
-           <Link href="/" className="opc">
-             Inicio
-           </Link>
-           <Link href="/Cuenta" className="opc btn--white prueba">
-             <span>Cuenta →</span>
-           </Link>
-         </>
+            <Link href="/Sobre_Nosotros" className="opc">
+              Sobre Nosotros
+            </Link>
+            <Link href="/Reportes" className="opc">
+              Reportes
+            </Link>
+            <Link href="/" className="opc">
+              Inicio
+            </Link>
+            <Link href="/Cuenta" className="opc btn--white prueba">
+              <span>Cuenta →</span>
+            </Link>
+          </>
         )}
       </div>
 
       {showConfirmation && (
-          <ConfirmationAlert className='alerta-custom-navbar'
-            message="¿Está seguro de que desea cerrar sesión?"
-            onConfirm={handleLogoutConfirmation}
-            onCancel={() => setShowConfirmation(false)}
-          />
+        <ConfirmationAlert
+          className="alerta-custom-navbar"
+          message="¿Está seguro de que desea cerrar sesión?"
+          onConfirm={handleLogoutConfirmation}
+          onCancel={() => setShowConfirmation(false)}
+        />
       )}
     </div>
   );
